requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
cfscrape==2.1.1
chardet==3.0.4
elasticsearch==7.9.1
emoji==0.6.0
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
openpyxl==3.0.5
pandas==1.1.3
psycopg2-binary==2.8.6
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-toolbelt==0.9.1
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.19
stem==1.8.0
torrequest==0.1.0
urllib3==1.25.10
xlrd==1.2.0
XlsxWriter==1.3.6
